requirements.txt
# Core FastAPI dependencies
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.7.1
pydantic-settings==2.4.0

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Database
supabase==2.17.0
asyncpg==0.29.0

# Vector embeddings
openai==1.3.6
tiktoken==0.5.2
numpy>=1.24.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Logging and monitoring
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Environment and configuration
python-dotenv==1.0.0

# Optional production enhancements
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4
# HTTP client for Cassidy integration
httpx
tenacity  # For retry logic with exponential backoff

# Testing
pytest
pytest-asyncio
pytest-mock
respx

# Development tools
black
isort
flake8

# Logging and monitoring
structlog
python-json-logger
